retraining_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.3
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
pandas==2.2.3
scikit-learn==1.5.2
mlflow==2.16.2
apscheduler==3.11.2
